Improved Offset Generation for Both Gyro & Accel
Gyro Calibration is now more accurate, My divisor was /8. I Changed it to /4. 
Note that the ITerm is a floating point number so dividing is more accurate than converting to integer and bit shifting >>3 or >>4.
Accel offset storage is also corrected to preserve Bit 0 Which is marked reserved in all the documentation. This could be the temperature compensation bit but I an finding little documentation about it. I didn't notice much improvement by preserving this bit with the MPU6050 but it isn't supposed to be changed so I'm preserving it.
NOTE: I tested the Accel Calibration on the MPU9250 (aka MPU6500 with a magnetometer) and failed to set any calibration values that made sense. Gurr! It seems that with the MPU9250, leaving the Accel Calibration at Factory Values and only Calibrating the Gyro provides the most accurate, stable Yaw Pitch and Roll Values. I am Searching for any hopefully working MPU9250 accelerometer calibration code to discover why my attempts are in vain.

+++ Arduino/MPU6050/MPU6050.cpp
@@ -3212,7 +3212,7 @@ void MPU6050::setDMPConfig2(uint8_t config) {
 /**
   @brief      Fully calibrate Gyro from ZERO in about 6-7 Loops 600-700 readings
 */
-void MPU6050::CalibrateGyro(int Loops = 6) {
+void MPU6050::CalibrateGyro(uint16_t Loops = 6) {
   int16_t Reading, Offset;
   float Error, PTerm, ITerm[3];
   double kP = 0.3;
